requirements/common.txt
Django==1.9.9
redis==2.10.5
hiredis==0.2.0
django-redis==4.4.4
requests==2.11.1
celery==3.1.23
django-celery==3.1.17
Pillow==3.3.1
django-extra-cbv==0.1.7.1
opengraph==0.5
scrapy==1.1.2
django-extensions==1.7.2
shortuuid==0.3.2
djangorestframework==3.4.6
validators==0.10.3

# tornado
TornadIO2==0.0.4
tornado==4.3
simplejson==3.8.2
tornado-redis==2.4.18
sockjs-tornado==1.0.3
-e git+https://github.com/evilkost/brukva.git#egg=brukva



